upgrade LuaJIT to 2.0.3

+++ scripting/lua/luajit/build_mac.sh
@@ -1,14 +1,16 @@
-#!/bin/sh
+#!/usr/bin/env bash
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
-SRCDIR=$DIR/LuaJit-2.0.1
-DESTDIR=$DIR/proj.mac
+SRCDIR=$DIR/LuaJit-2.0.3
+DESTDIR=$DIR/mac
 MACOSX_DEPLOYMENT_TARGET="10.6"
 
-rm "$DESTDIR"/*.a
+if [ -f "$DESTDIR"/*.a ]; then
+    rm "$DESTDIR"/*.a
+fi;
 
 cd $SRCDIR
 make clean
-make CC="gcc -m32 -arch i386" clean all
+make CC="clang -arch x86_64" clean all
 
 if [ -f $SRCDIR/src/libluajit.a ]; then
     mv $SRCDIR/src/libluajit.a $DESTDIR/libluajit2.a
